editing: extract ObjectEditService, reducing the controller to a facade
Final step of the ObjectCommandController split: object placement, deletion,
move, rotation, FRM change, exit-grid creation/edit, and per-instance editing —
plus the ExitGridCommandState / MapObjectInstanceState value types — move into
ObjectEditService, built on the shared UndoBatcher.

ObjectCommandController.cpp drops from ~940 to ~170 lines: it now just constructs
the six editing services and forwards each public method to the right one. The
move/rotation undo+redo halves, which duplicated their apply loops, now share
applyObjectPositions()/applyObjectDirections() helpers (also bringing both methods
back under the complexity limit). Unused includes pruned. 235 tests pass.
