+++ src/ui/editing/ObjectCommandController.cpp
@@ -1,17 +1,8 @@
 #include "ObjectCommandController.h"
 
-#include "editor/HexagonGrid.h"
-#include "editor/Object.h"
-#include "format/map/Map.h"
-#include "format/map/MapObject.h"
-#include "format/map/Tile.h"
-#include "resource/GameResources.h"
-#include "editor/TileChange.h"
-#include "util/UndoStack.h"
-#include "ui/rendering/MapSpriteLoader.h"
-
-#include <algorithm>
-#include <spdlog/spdlog.h>
+// ObjectCommandController is a thin facade: its constructor wires up the editing
+// services and every method delegates to one of them, so only the service/param
+// types (forward-declared via the header) are needed here.
 
 namespace geck {
 
@@ -40,7 +31,8 @@ ObjectCommandController::ObjectCommandController(resource::GameResources& resour
     , _scriptService(map, _batcher)
     , _refreshObjects(std::move(refreshObjects))
     , _reloadTiles(std::move(reloadTiles))
-    , _mapService(map, objects, wallBlockerOverlays, _scriptService, _batcher, _refreshObjects, _reloadTiles) {
+    , _mapService(map, objects, wallBlockerOverlays, _scriptService, _batcher, _refreshObjects, _reloadTiles)
+    , _objectService(resources, map, hexgrid, mapSpriteLoader, objects, wallBlockerOverlays, _batcher, _refreshObjects) {
 }
 
 void ObjectCommandController::applyTileChanges(const std::vector<TileChange>& changes, bool applyAfterState) {
@@ -56,259 +48,47 @@ void ObjectCommandController::applyTileEdit(const std::string& description, cons
 }
 
 void ObjectCommandController::addObjectData(const std::shared_ptr<MapObject>& mapObject) {
-    if (!mapObject || !_map) {
-        spdlog::warn("addObjectData: Invalid parameters - mapObject:{} map:{}",
-            mapObject != nullptr, _map != nullptr);
-        return;
-    }
-    if (mapObject->elevation >= 3) {
-        spdlog::error("addObjectData: Invalid elevation {}", mapObject->elevation);
-        return;
-    }
-    _map->getMapFile().map_objects[mapObject->elevation].push_back(mapObject);
+    _objectService.addObjectData(mapObject);
 }
 
 void ObjectCommandController::removeObjectData(const std::shared_ptr<MapObject>& mapObject) {
-    if (!mapObject || !_map) {
-        spdlog::warn("removeObjectData: Invalid parameters - mapObject:{} map:{}",
-            mapObject != nullptr, _map != nullptr);
-        return;
-    }
-    if (mapObject->elevation >= 3) {
-        spdlog::error("removeObjectData: Invalid elevation {}", mapObject->elevation);
-        return;
-    }
-    auto& elevationObjects = _map->getMapFile().map_objects[mapObject->elevation];
-    std::erase(elevationObjects, mapObject);
-    _refreshObjects();
+    _objectService.removeObjectData(mapObject);
 }
 
 void ObjectCommandController::addPlacedObject(const std::shared_ptr<MapObject>& mapObject, const std::shared_ptr<Object>& object) {
-    if (!mapObject || !object || !_map) {
-        spdlog::warn("addPlacedObject: Invalid parameters - mapObject:{} object:{} map:{}",
-            mapObject != nullptr, object != nullptr, _map != nullptr);
-        return;
-    }
-    if (mapObject->elevation >= 3) {
-        spdlog::error("addPlacedObject: Invalid elevation {}", mapObject->elevation);
-        return;
-    }
-
-    auto& mapFile = _map->getMapFile();
-    mapFile.map_objects[mapObject->elevation].push_back(mapObject);
-    _mapSpriteLoader.appendObjectSprite(mapObject, object, _objects, _wallBlockerOverlays);
+    _objectService.addPlacedObject(mapObject, object);
 }
 
 void ObjectCommandController::removePlacedObject(const std::shared_ptr<MapObject>& mapObject, const std::shared_ptr<Object>& object) {
-    if (!mapObject || !_map) {
-        spdlog::warn("removePlacedObject: Invalid parameters - mapObject:{} map:{}",
-            mapObject != nullptr, _map != nullptr);
-        return;
-    }
-    if (mapObject->elevation >= 3) {
-        spdlog::error("removePlacedObject: Invalid elevation {}", mapObject->elevation);
-        return;
-    }
-
-    auto& mapFile = _map->getMapFile();
-    auto& elevationObjects = mapFile.map_objects[mapObject->elevation];
-    elevationObjects.erase(std::remove(elevationObjects.begin(), elevationObjects.end(), mapObject), elevationObjects.end());
-    if (object) {
-        _objects.erase(std::remove(_objects.begin(), _objects.end(), object), _objects.end());
-    }
-    _refreshObjects();
+    _objectService.removePlacedObject(mapObject, object);
 }
 
 bool ObjectCommandController::registerObjectPlacement(const std::shared_ptr<MapObject>& mapObject, const std::shared_ptr<Object>& object) {
-    if (!mapObject || !object) {
-        spdlog::warn("registerObjectPlacement: Invalid parameters - mapObject:{} object:{}",
-            mapObject != nullptr, object != nullptr);
-        return false;
-    }
-
-    UndoCommand cmd;
-    cmd.description = "Place Object";
-    cmd.redo = [this, mapObject, object]() {
-        if (mapObject && object) {
-            addPlacedObject(mapObject, object);
-        }
-    };
-    cmd.undo = [this, mapObject, object]() {
-        if (mapObject && object) {
-            removePlacedObject(mapObject, object);
-        }
-    };
-
-    cmd.redo();
-    return pushCommand(std::move(cmd));
+    return _objectService.registerObjectPlacement(mapObject, object);
 }
 
 bool ObjectCommandController::registerObjectData(const std::shared_ptr<MapObject>& mapObject) {
-    if (!mapObject) {
-        spdlog::warn("registerObjectData: null mapObject");
-        return false;
-    }
-
-    UndoCommand cmd;
-    cmd.description = "Place Object";
-    cmd.redo = [this, mapObject]() {
-        if (mapObject) {
-            addObjectData(mapObject);
-        }
-    };
-    cmd.undo = [this, mapObject]() {
-        if (mapObject) {
-            removeObjectData(mapObject);
-        }
-    };
-
-    cmd.redo();
-    return pushCommand(std::move(cmd));
+    return _objectService.registerObjectData(mapObject);
 }
 
 bool ObjectCommandController::registerObjectDeletion(
     const std::vector<std::pair<std::shared_ptr<MapObject>, std::shared_ptr<Object>>>& removedObjects) {
-    if (removedObjects.empty()) {
-        return false;
-    }
-
-    // The objects are already removed by the caller; this only records undo/redo.
-    UndoCommand cmd;
-    cmd.description = "Delete Objects";
-    cmd.undo = [this, removedObjects]() {
-        for (const auto& pair : removedObjects) {
-            if (pair.first && pair.second) {
-                addPlacedObject(pair.first, pair.second);
-            }
-        }
-    };
-    cmd.redo = [this, removedObjects]() {
-        for (const auto& pair : removedObjects) {
-            if (pair.first && pair.second) {
-                removePlacedObject(pair.first, pair.second);
-            }
-        }
-    };
-    return pushCommand(std::move(cmd));
+    return _objectService.registerObjectDeletion(removedObjects);
 }
 
 bool ObjectCommandController::registerObjectMove(const std::vector<std::shared_ptr<Object>>& objects,
     const std::vector<std::pair<int, int>>& moves) {
-    if (objects.empty() || moves.empty() || objects.size() != moves.size()) {
-        spdlog::warn("registerObjectMove: Invalid input - objects:{} moves:{}", objects.size(), moves.size());
-        return false;
-    }
-
-    UndoCommand cmd;
-    cmd.description = "Move Objects";
-    cmd.undo = [this, objects, moves]() {
-        if (_hexgrid.empty()) {
-            spdlog::error("registerObjectMove undo: hex grid not available");
-            return;
-        }
-        for (size_t i = 0; i < objects.size() && i < moves.size(); ++i) {
-            auto object = objects[i];
-            if (!object || !object->hasMapObject()) {
-                continue;
-            }
-
-            auto& mapObj = object->getMapObject();
-            mapObj.position = moves[i].first;
-            if (auto hex = _hexgrid.getHexByPosition(static_cast<uint32_t>(moves[i].first)); hex.has_value()) {
-                object->setHexPosition(hex->get());
-            } else {
-                spdlog::warn("registerObjectMove undo: Invalid hex position {}", moves[i].first);
-            }
-        }
-        _refreshObjects();
-    };
-    cmd.redo = [this, objects, moves]() {
-        if (_hexgrid.empty()) {
-            spdlog::error("registerObjectMove redo: hex grid not available");
-            return;
-        }
-        for (size_t i = 0; i < objects.size() && i < moves.size(); ++i) {
-            auto object = objects[i];
-            if (!object || !object->hasMapObject()) {
-                continue;
-            }
-
-            auto& mapObj = object->getMapObject();
-            mapObj.position = moves[i].second;
-            if (auto hex = _hexgrid.getHexByPosition(static_cast<uint32_t>(moves[i].second)); hex.has_value()) {
-                object->setHexPosition(hex->get());
-            } else {
-                spdlog::warn("registerObjectMove redo: Invalid hex position {}", moves[i].second);
-            }
-        }
-        _refreshObjects();
-    };
-
-    cmd.redo();
-    return pushCommand(std::move(cmd));
+    return _objectService.registerObjectMove(objects, moves);
 }
 
 bool ObjectCommandController::registerObjectRotation(const std::vector<std::shared_ptr<Object>>& objects,
     const std::vector<int>& beforeDirs,
     const std::vector<int>& afterDirs) {
-    if (objects.empty() || beforeDirs.size() != objects.size() || afterDirs.size() != objects.size()) {
-        spdlog::warn("registerObjectRotation: Invalid input sizes - objects:{} before:{} after:{}",
-            objects.size(), beforeDirs.size(), afterDirs.size());
-        return false;
-    }
-
-    UndoCommand cmd;
-    cmd.description = "Rotate Objects";
-    cmd.undo = [objects, beforeDirs]() {
-        for (size_t i = 0; i < objects.size() && i < beforeDirs.size(); ++i) {
-            if (!objects[i]) {
-                continue;
-            }
-            if (beforeDirs[i] < 0 || beforeDirs[i] > 5) {
-                spdlog::warn("registerObjectRotation undo: Invalid direction {}", beforeDirs[i]);
-                continue;
-            }
-            objects[i]->setDirection(static_cast<ObjectDirection>(beforeDirs[i]));
-        }
-    };
-    cmd.redo = [objects, afterDirs]() {
-        for (size_t i = 0; i < objects.size() && i < afterDirs.size(); ++i) {
-            if (!objects[i]) {
-                continue;
-            }
-            if (afterDirs[i] < 0 || afterDirs[i] > 5) {
-                spdlog::warn("registerObjectRotation redo: Invalid direction {}", afterDirs[i]);
-                continue;
-            }
-            objects[i]->setDirection(static_cast<ObjectDirection>(afterDirs[i]));
-        }
-    };
-
-    return pushCommand(std::move(cmd));
+    return _objectService.registerObjectRotation(objects, beforeDirs, afterDirs);
 }
 
 void ObjectCommandController::applyFrmToObject(const std::shared_ptr<Object>& object, uint32_t frmPid, const std::string& frmPath) {
-    if (!object) {
-        spdlog::warn("applyFrmToObject: null object provided");
-        return;
-    }
-    if (frmPath.empty()) {
-        spdlog::warn("applyFrmToObject: empty FRM path provided");
-        return;
-    }
-
-    try {
-        const auto& texture = _resources.textures().get(frmPath);
-        sf::Sprite sprite(texture);
-        sprite.setPosition(object->getSprite().getPosition());
-        object->setSprite(std::move(sprite));
-        if (object->hasMapObject()) {
-            object->getMapObject().frm_pid = frmPid;
-            object->setDirection(static_cast<ObjectDirection>(object->getMapObject().direction));
-        }
-    } catch (const std::exception& e) {
-        spdlog::error("ObjectCommandController::applyFrmToObject - failed to apply FRM {}: {}", frmPath, e.what());
-    }
+    _objectService.applyFrmToObject(object, frmPid, frmPath);
 }
 
 bool ObjectCommandController::registerObjectFrmChange(const std::shared_ptr<Object>& object,
@@ -316,169 +96,28 @@ bool ObjectCommandController::registerObjectFrmChange(const std::shared_ptr<Obje
     const std::string& oldFrmPath,
     uint32_t newFrmPid,
     const std::string& newFrmPath) {
-    if (!object) {
-        return false;
-    }
-    if (newFrmPath.empty()) {
-        spdlog::warn("registerObjectFrmChange: new FRM path empty, ignoring");
-        return false;
-    }
-
-    UndoCommand cmd;
-    cmd.description = "Change Object FRM";
-    cmd.undo = [this, object, oldFrmPid, oldFrmPath]() {
-        applyFrmToObject(object, oldFrmPid, oldFrmPath);
-    };
-    cmd.redo = [this, object, newFrmPid, newFrmPath]() {
-        applyFrmToObject(object, newFrmPid, newFrmPath);
-    };
-
-    cmd.redo();
-    return pushCommand(std::move(cmd));
+    return _objectService.registerObjectFrmChange(object, oldFrmPid, oldFrmPath, newFrmPid, newFrmPath);
 }
 
 bool ObjectCommandController::registerExitGridCreation(const std::vector<std::shared_ptr<MapObject>>& exitGrids, int elevation) {
-    if (exitGrids.empty()) {
-        return false;
-    }
-
-    UndoCommand cmd;
-    cmd.description = exitGrids.size() == 1 ? "Place Exit Grid" : "Place Exit Grids";
-    cmd.undo = [this, exitGrids, elevation]() {
-        if (!_map) {
-            return;
-        }
-
-        auto& mapFile = _map->getMapFile();
-        if (mapFile.map_objects.find(elevation) == mapFile.map_objects.end()) {
-            return;
-        }
-
-        auto& elevationObjects = mapFile.map_objects[elevation];
-        for (const auto& exitGrid : exitGrids) {
-            elevationObjects.erase(
-                std::remove(elevationObjects.begin(), elevationObjects.end(), exitGrid),
-                elevationObjects.end());
-        }
-        _refreshObjects();
-    };
-    cmd.redo = [this, exitGrids, elevation]() {
-        if (!_map) {
-            return;
-        }
-
-        auto& mapFile = _map->getMapFile();
-        for (const auto& exitGrid : exitGrids) {
-            mapFile.map_objects[elevation].push_back(exitGrid);
-        }
-        _refreshObjects();
-    };
-
-    cmd.redo();
-    return pushCommand(std::move(cmd));
+    return _objectService.registerExitGridCreation(exitGrids, elevation);
 }
 
 bool ObjectCommandController::registerExitGridEdit(const std::vector<std::shared_ptr<MapObject>>& exitGrids,
     const std::vector<ExitGridCommandState>& beforeStates,
     const std::vector<ExitGridCommandState>& afterStates) {
-    if (exitGrids.empty() || exitGrids.size() != beforeStates.size() || exitGrids.size() != afterStates.size()) {
-        spdlog::warn("registerExitGridEdit: Invalid input sizes - grids:{} before:{} after:{}",
-            exitGrids.size(), beforeStates.size(), afterStates.size());
-        return false;
-    }
-
-    UndoCommand cmd;
-    cmd.description = exitGrids.size() == 1 ? "Edit Exit Grid" : "Edit Exit Grids";
-    cmd.undo = [this, exitGrids, beforeStates]() {
-        for (size_t i = 0; i < exitGrids.size(); ++i) {
-            if (!exitGrids[i]) {
-                continue;
-            }
-            exitGrids[i]->exit_map = beforeStates[i].exitMap;
-            exitGrids[i]->exit_position = beforeStates[i].exitPosition;
-            exitGrids[i]->exit_elevation = beforeStates[i].exitElevation;
-            exitGrids[i]->exit_orientation = beforeStates[i].exitOrientation;
-            exitGrids[i]->frm_pid = beforeStates[i].frmPid;
-            exitGrids[i]->pro_pid = beforeStates[i].proPid;
-        }
-        _refreshObjects();
-    };
-    cmd.redo = [this, exitGrids, afterStates]() {
-        for (size_t i = 0; i < exitGrids.size(); ++i) {
-            if (!exitGrids[i]) {
-                continue;
-            }
-            exitGrids[i]->exit_map = afterStates[i].exitMap;
-            exitGrids[i]->exit_position = afterStates[i].exitPosition;
-            exitGrids[i]->exit_elevation = afterStates[i].exitElevation;
-            exitGrids[i]->exit_orientation = afterStates[i].exitOrientation;
-            exitGrids[i]->frm_pid = afterStates[i].frmPid;
-            exitGrids[i]->pro_pid = afterStates[i].proPid;
-        }
-        _refreshObjects();
-    };
-
-    return pushCommand(std::move(cmd));
+    return _objectService.registerExitGridEdit(exitGrids, beforeStates, afterStates);
 }
 
-MapObjectInstanceState ObjectCommandController::captureInstanceState(const MapObject& o) {
-    MapObjectInstanceState s;
-    s.flags = o.flags;
-    s.dataFlags = o.unknown11;
-    s.lightRadius = o.light_radius;
-    s.lightIntensity = o.light_intensity;
-    s.walkthrough = o.walkthrough;
-    s.map = o.map;
-    s.elevhex = o.elevhex;
-    s.elevtype = o.elevtype;
-    s.elevlevel = o.elevlevel;
-    s.aiPacket = o.ai_packet;
-    s.groupId = o.group_id;
-    s.currentHp = o.current_hp;
-    s.currentRad = o.current_rad;
-    s.currentPoison = o.current_poison;
-    return s;
-}
-
-void ObjectCommandController::applyInstanceState(MapObject& o, const MapObjectInstanceState& s) {
-    o.flags = s.flags;
-    o.unknown11 = s.dataFlags;
-    o.light_radius = s.lightRadius;
-    o.light_intensity = s.lightIntensity;
-    o.walkthrough = s.walkthrough;
-    o.map = s.map;
-    o.elevhex = s.elevhex;
-    o.elevtype = s.elevtype;
-    o.elevlevel = s.elevlevel;
-    o.ai_packet = s.aiPacket;
-    o.group_id = s.groupId;
-    o.current_hp = s.currentHp;
-    o.current_rad = s.currentRad;
-    o.current_poison = s.currentPoison;
+MapObjectInstanceState ObjectCommandController::captureInstanceState(const MapObject& object) {
+    return ObjectEditService::captureInstanceState(object);
 }
 
 bool ObjectCommandController::registerInstanceEdit(const std::shared_ptr<MapObject>& mapObject,
     const MapObjectInstanceState& before,
     const MapObjectInstanceState& after,
     const std::string& description) {
-    if (!mapObject) {
-        spdlog::warn("registerInstanceEdit: null mapObject");
-        return false;
-    }
-
-    UndoCommand cmd;
-    cmd.description = description;
-    cmd.undo = [this, mapObject, before]() {
-        applyInstanceState(*mapObject, before);
-        _refreshObjects();
-    };
-    cmd.redo = [this, mapObject, after]() {
-        applyInstanceState(*mapObject, after);
-        _refreshObjects();
-    };
-
-    cmd.redo();
-    return pushCommand(std::move(cmd));
+    return _objectService.registerInstanceEdit(mapObject, before, after, description);
 }
 
 std::vector<std::shared_ptr<MapObject>> ObjectCommandController::cloneInventory(
